ENH: Allow Dataset.to_cloud to infer cloud location
